+++ compiler/src/jdk.graal.compiler/src/jdk/graal/compiler/core/phases/HighTier.java
@@ -124,6 +124,10 @@ public class HighTier extends BaseTier<HighTierContext> {
             appendPhase(new LoopFullUnrollPhase(canonicalizer, loopPolicies));
         }
 
+        if (GraalOptions.OptReadElimination.getValue(options)) {
+            appendPhase(new ReadEliminationPhase(canonicalizer));
+        }
+
         if (GraalOptions.LoopPeeling.getValue(options)) {
             appendPhase(new LoopPeelingPhase(loopPolicies, canonicalizer));
         }